Add permission management to Auth service
Introduces permission entities, repositories, and service interfaces to MyApp.Auth for user and role-based permission checks. Adds PermissionsController API endpoint, implements repository and service logic, and updates all projects to .NET 9 and latest package versions. Also adds shared permission checker interfaces and attributes in MyApp.Shared.

+++ MyApp.Sales/MyApp.Sales.Application/Mapping/SalesOrderMappingProfile.cs
@@ -9,7 +9,7 @@ namespace MyApp.Sales.Application.Mapping
         public SalesOrderMappingProfile()
         {
             CreateMap<SalesOrder, SalesOrderDto>()
-                .ForMember(d => d.Status, opt => opt.MapFrom(s => (int)s.Status))
+                .ForMember(d => d.Status, opt => opt.MapFrom(s => s.Status))
                 .ReverseMap()
                 .ForMember(d => d.Status, opt => opt.MapFrom(s => (SalesOrderStatus)s.Status));
